requirements.txt
torch
transformers
pillow
numpy
diffusers
accelerate
safetensors